% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates a visual Map of Accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number corresponding to a US state based on FARS data.}

\item{year}{A year for which to summarize data.}
}
\value{
Returns a map plot of the US with accidents for a specified state
and year.  Returns an error if the state or year provided are invalid.
}
\description{
Creates a visual Map of Accident data for specified year \code{year}
occuring in a specified state \code{state.num}.
}
\examples{
\dontrun{
fars_map_state(1,2013)
}

}
