% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read}
\alias{fars_read}
\title{Reads a CSV file into a data frame.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{File path of a CSV file (character)}
}
\value{
Returns a (data.frame) based on the CSV file.
}
\description{
Reads a CSV file from \code{filename} using readr::read_csv and converts
the resulting tibble into a data frame.
}
\examples{
\dontrun{
accident_2015 <- fars_read("data/accident_2013.csv.bz2")
}

}
