% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_read_years}
\alias{fars_read_years}
\title{Get Months and Year from Accident Data Files}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{Years of data to be retrieved.}
}
\value{
Returns a list of (data.frames) representing the month and year
data from available Accident data files.  Provides and error message if
years requested are not available.
}
\description{
Extracts Month and Year data from Accident datafiles for specified years.
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years("2013")
fars_read_years(2013:2015)
}

}
