% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Calculates a summary of accidents per month for given years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Years of data to be summarized.}
}
\value{
Returns a (data.frame) with number of accidents per month for
the specified years.

importFrom dplyr bind_rows group_by summarize
importFrom tidyr spread
}
\description{
Accepts a list of years \code{years} and provides a summary of total
accidents for each month in those years.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years("2013")
fars_summarize_years(2013:2015)
}

}
