% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.r
\name{make_filename}
\alias{make_filename}
\title{Create filename with variable year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Specified year of data to be used in the filename.}
}
\value{
This function returns a filename to be used for a specific year
of accident data.
}
\description{
Create a filename in the format "accident_XXXX.csv.bz2 where XXXX is
replaced by a year supplied to the function.
}
\examples{
\dontrun{
make_filename(2013)
make_filename("2013")
}

}
